requirements.txt
appdirs==1.4.4
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.2
cssselect==1.2.0
defusedxml==0.7.1
et-xmlfile==1.1.0
fake-useragent==1.3.0
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
lxml==4.9.3
mysql-connector-python==8.2.0
numpy==1.26.2
odfpy==1.4.1
openpyxl==3.1.2
outcome==1.3.0.post0
pandas==2.1.3
parse==1.19.1
protobuf==4.21.12
pyee==8.2.2
pyppeteer==1.0.2
pyquery==2.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
schedule==1.2.1
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tqdm==4.66.1
trio==0.23.1
trio-websocket==0.11.1
tzdata==2023.3
urllib3==1.26.18
w3lib==2.1.2
websockets==10.4
wsproto==1.2.0
zipp==3.17.0
